# WIP - Optimize compile time by changing datastructures