--- a/lib/compiler/src/Makefile
+++ b/lib/compiler/src/Makefile
@@ -70,6 +70,7 @@ MODULES =  \
 	cerl \
 	cerl_clauses \
 	cerl_inline \
+	cerl_sets \
 	cerl_trees \
 	compile \
 	core_lib \