… bootstrap/lib/compiler/ebin/beam_dead.beam
… bootstrap/lib/compiler/ebin/beam_dict.beam
… bootstrap/lib/compiler/ebin/beam_jump.beam
… bootstrap/lib/compiler/ebin/beam_type.beam
… bootstrap/lib/compiler/ebin/cerl_sets.beam
… bootstrap/lib/compiler/ebin/sys_core_fold.beam
… bootstrap/lib/compiler/ebin/v3_codegen.beam
… bootstrap/lib/compiler/ebin/v3_kernel.beam
… bootstrap/lib/kernel/ebin/inet_dns.beam
… bootstrap/lib/kernel/ebin/inet_tcp_dist.beam
… bootstrap/lib/stdlib/ebin/c.beam
… bootstrap/lib/stdlib/ebin/erl_anno.beam
… bootstrap/lib/stdlib/ebin/erl_lint.beam
… bootstrap/lib/stdlib/ebin/shell_default.beam
… lib/compiler/src/v3_codegen.erl
… lib/compiler/src/v3_kernel.erl
… lib/compiler/src/v3_life.erl