Change some of the data structures to Maps to gain some speed.

Preliminary findings using `time erlc $testfile` and best of three:

`$testfile` |  `OTP-18.0-rc2-1-g9a81b28` |  `OTP-18.0-rc2-18-g43a4893`
-----|---|----
NBAP-PDU-Contents.asn |  26146 ms  | 22810 ms
erl_parse.erl | 1591 ms | 1439 ms
lists.erl  | 579 ms  | 518 ms

This is about ~10% improvement.