@josevalim 2015-05-16
I was literally trying this in some places now. :D You probably know this but in Elixir we also got better performance in some cases by replacing sets by maps where the value is `true`.