@essen 2015-05-16
@psyeugenic Here is another file you may find interesting for optimizing the compiler: https://github.com/ninenines/cowlib/blob/master/src/cow_http_hd.erl

It has two include files required: https://github.com/ninenines/cowlib/tree/master/include

In this one's case it's mostly huge guards. It takes a few seconds to compile on my machine currently. Feel free to take a stab at it. :-)