@kostis 2015-05-16
A word of caution: gb_sets and gb_trees are data structures that are "optimized for Big-O complexity", i.e., they give guaranteed O(log(n)) times for insertion and lookup at the expense of their operations being heavier.

Asking out of curiosity, are maps in 18.0 efficient enough to guarantee O(log(n)) complexity, or better, for insertion and lookup, or are there still cases where O(n) operations are taking place?