@psyeugenic 2015-05-16
@kostis That is a reasonable concern. Maps has a time complexity of O(log(n)) for inserts and lookups and has a very low constant before that O. The Maps tree is probabilistically balanced, i.e. it is a function of the hash-function so for a reasonable large set of data you will, by law of distribution, have reasonable balanced tree. Assuming a good hash-function.

A couple of things to note. Maps has a shallower tree than gb_trees, which means lookups are much faster in Maps. There is only one key compare per lookup in Maps versus one at each level for gb_trees. The other thing is, for few elements in a Map, i.e. fewer than 33, we do linear searches. This is usually pretty fast though due to locality but 32 elements is on the upper limit .. I would like to lower it. That's my concern at least.

There are ofc cases where Maps would be less ideal than say gb_trees but I would say Maps is an overall improvement in this particular case. And of course, if you need an ordered set then gb_trees is what you want in most cases. It depends on the situation.