@psyeugenic 2015-05-17
Preliminary findings using `time erlc $testfile` and best of three:

`$testfile` |  `OTP-18.0-rc2-1-g9a81b28` |  `OTP-18.0-rc2-19-g4c9459a`
-----|---|----
NBAP-PDU-Contents.asn |  26146 ms  | 22668 ms
erl_parse.erl | 1591 ms | 1436 ms
lists.erl  | 579 ms  | 513 ms
cow_http_hd.erl | 7530 ms | 6511 ms

There is at least one other bottleneck that irritates me and that's `v3_lifes` Variable lifetime mapping used by `v3_codegen`. However I think we'll save that one for when we remove the `v3_life` pass. What little is done it that pass should be move to `v3_codegen` and/or `v3_kernel`.