@josevalim 2015-05-18
I can confirm those results too. I merged this with my other patches and compiling Elixir (both erl+elixir sources) is about 15% faster than Erlang 17. I will continue using those patches to see if anything weird pops up.