@OTP-Maintainer 2015-05-20
The summary line of the commit message is too long and/or ends with a "."
Make sure the whole message follows the guidelines here: https://github.com/erlang/otp/wiki/Writing-good-commit-messages.

Bad message: Merge branch 'hans/ssh/ssh_msg_debug_fun/OTP-12738' into maint-17
Merge branch 'hans/inets/bad_content_length/OTP-12739' into maint-17
ssh: option for handling the SSH_MSG_DEBUG message's printouts


----------------

I am a script, I am not human 

----------------