@psyeugenic 2015-05-21
Moved (merged) `maps:filter/2` to master and rebased this branch to current master.